refactor(email): split dictionary email resolution helpers by responsibility
- extract domain candidate + fuzzy-domain ranking logic into `DictionaryMatcher+EmailDomainResolution.swift`
- extract shared local/domain normalization and attached-list-marker parsing into `DictionaryMatcher+EmailParsing.swift`
- keep `DictionaryMatcher+EmailResolution.swift` focused on spoken/literal/standalone resolution paths
- wire new files into `KeyVox.xcodeproj/project.pbxproj`
- update `Docs/CODEMAP.md` and `Docs/ENGINEERING.md` to reflect the new email helper structure
